backend/requirements.txt
fastapi
uvicorn[standard]
httpx
pydantic-settings
python-dotenv
sqlalchemy[asyncio]
asyncpg
alembic
PyJWT
cryptography
